Add slots to blocks
Each block gets a new slot each time it is entered in a machine state.
This slot is then "current" for the block and the state.

A slot is a Value_Table, so same backing store as 'value table' objects.

The idea is that one can use slots to capture "arguments" to blocks.