Cloning of values
Cloning is the act of creating a deep copy of a value, instead of a new
reference to the same value. Doing this is more expensive in terms of
time and memory, but is in general transparent to the running program,
as values are meant to be immutable.

The use of this is to allow appending data to tuples and tables by
returning an entirely new object without altering the original one.