Add a ValueLoader object
This class has the same underlying functionality as the OperationLoader, but
it handles the logic for Value. The main difference is that operations use
an integer ID as the "key" in the binary format, but Values uses a per-class MARKER byte.

This is preserved in the new API in order to keep the serialization format stable.
I will probably go back and revisit that at some point, but not right now.

Doing this has the main purpose of allowing value types to be identifiers instead of
magic keywords, just like for Operations.

A few error messages change here and there, but otherwise this has very little impact
on tests and general system behavior.