Initial definition of a Value_Callable
Value_Callable shall be the fundamental currency of "things one can execute", so
 - operations;
 - blocks;
 - binds.

It will eventually replace Value_{Operation|Block|Bind} and become the common value type of all things executable.

This sketched out implementation does not involve any testing, as Value_Callable is yet to be bound to anything else in the system.
It does however come with an initial implementation of parsing a Callable.