Sketch out a select operation
Select picks the topmost value of the stack and compares it to every entry in a table.
If an entry matches, the value of that entry is executed (if an operation).
If the value is not an operation, or no entry matches, then an optional "else" entry is executed.

The "else" defaults to nop.

Test cases will be pushed in a subsequent commit