Introduce a CALL operation
call is a convenience for pushing a set of values on the stack and loading and executing a named operation

This implementation is such that the order of arguments listed in the call tuple is the opposite of
what gets pushed out to the slots of the called blocks, if any exist.

This is probably suboptimal and should be fixed by pushing in the reverse order, such that one can read
the call and expect the slots to be filled properly.