Introduce NativeOperations support
NativeOperations is a container (one per MachineState) of Buckets.
Each Bucket has a name, and contains 0 or more NativeOperationLoaders, which are themselves a pair of functions:
  - one from Parser to Native operation;
  - one from ByteStream to Native operation.

The goal is for buckets to function as very loose "library" objects, e.g. time support, file support, network, ....
each of which contains the native operations for that bucket.