Rework of the way native operations load themselves
Native libraries provide a descriptor, which is the combination of a bucket name and a list of operation loaders;
Operation loaders take a bucket and make a new Native.

This allows one to go through the loaders, initialize them with a bucket's qualified name, and obtain the Value_Operation objects that end up in the value_store.

There are only a few minor bits and pieces missing, but this is largely a sustainable long-term model for native operations.