Give BaseOperation support for preconditions
A BaseOperation can include a default-constructible Preconditions object, which can return an ErrorCode,
causing the entire operation to fail before executing any op-specific code.

This can be used to provide off-the-shelf validation, e.g. argument counts and types

This is used in providing argument count validation for most operations in a centralized way, saving more than 40 lines of code in the process