Add a value type for ATOM
ATOM is any valid identifier stored as a string. It can be compared for equality to other atoms and be selected against

In itself, it's not much more useful than a string, but the goal is to use atoms as the backbone for ADT, such that one could describe an ADT as a set of ATOMs
For instance, boolean would become oneof(set[atom true, atom false]) for some appropriate syntax.